Se soluciona el bug de la importacion en el guardian del token.
