+++ src/core/auth/guards/roles.guard.ts
@@ -1,11 +1,34 @@
-import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
+import {
+  CanActivate,
+  ExecutionContext,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
+import { Reflector } from '@nestjs/core';
 import { Observable } from 'rxjs';
+
+import { ROLES_KEY } from '../decorators/roles.decorator';
+import { Role } from '../interface/IRole.interface';
+import { IPayloadToken } from '../interface/IPayloadToken.interface';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
+  constructor(private reflector: Reflector) {}
   canActivate(
     context: ExecutionContext,
   ): boolean | Promise<boolean> | Observable<boolean> {
-    return true;
+    const roles = this.reflector.get<Role[]>(ROLES_KEY, context.getHandler());
+    if (!roles) {
+      return true;
+    }
+
+    const request = context.switchToHttp().getRequest();
+    const user = request.user as IPayloadToken;
+
+    const isAuth = roles.some((role) => role === user.role);
+    if (!isAuth) {
+      throw new UnauthorizedException('your role is wrong');
+    }
+    return isAuth;
   }
 }
