+++ src/core/auth/strategies/jwt.strategy.ts
@@ -1,11 +1,10 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
-import { Strategy } from 'passport-local';
-import { ExtractJwt } from 'passport-jwt';
+import { ExtractJwt, Strategy } from 'passport-jwt';
+import { ConfigType } from '@nestjs/config';
 
 import { IPayloadToken } from '../interface/IPayloadToken.interface';
 import config from 'src/config';
-import { ConfigType } from '@nestjs/config';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
